refactor(charts): split TokenHistoryChart into modular components
Split TokenHistoryChart.tsx (991 lines) into focused, maintainable modules:

- types.ts: Shared type definitions
- utils.ts: Shared formatting and utility functions
- components.tsx: Shared UI components (LegendItem, CustomTooltip, ChartWrapper)
- DailyChart.tsx: Daily bar chart component
- HourlyChart.tsx: Hourly area chart component
- index.ts: Public API exports

Benefits:
- Each chart type is now independently maintainable
- Shared utilities eliminate duplication
- Easier to add new chart types or modify existing ones
- Main component reduced to simple wrapper/delegation logic
