+++ frontend/src/components/dashboard/TokenHistoryChart/index.ts
@@ -1,0 +1,20 @@
+// TokenHistoryChart component exports
+
+export { DailyTokenHistoryChart } from './DailyChart';
+export { HourlyTokenHistoryChart } from './HourlyChart';
+export { TokenHistoryChart } from './TokenHistoryChart';
+export { ChartWrapper, LegendItem, CustomTooltip } from './components';
+export {
+    formatTimeLabel,
+    formatTooltipTime,
+    formatChartData,
+    calculateLabelInterval,
+    formatYAxis,
+    formatTooltipValue,
+} from './utils';
+export type {
+    TimeSeriesData,
+    ChartDataPoint,
+    SeriesVisibility,
+    LegendItemProps,
+} from './types';
